feat: rename VaTimePickerColumn directory
