+++ packages/ui/src/components/va-time-picker/VaTimePickerColumn/_variables.scss
